update :: SimpleDocumentResponse.java

+++ daemawiki-domain/src/main/java/org/daemawiki/domain/document/dto/response/SimpleDocumentResponse.java
@@ -10,10 +10,10 @@ import java.time.LocalDateTime;
 public record SimpleDocumentResponse(
         String id,
         String title,
-        Integer numberOfUpdate,
+        Long numberOfUpdate,
         @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", locale = "Asia/Seoul")
         LocalDateTime updatedDate,
-        Integer view
+        Long view
 ) {
 
         public static SimpleDocumentResponse of(DefaultDocument document) {
